+++ infra/terraform/aws-storage/iam.tf
@@ -17,7 +17,6 @@ resource "aws_iam_user_policy" "storage" {
           "s3:PutObject",
           "s3:GetObject",
           "s3:DeleteObject",
-          "s3:HeadObject",
         ]
         Resource = "${aws_s3_bucket.files.arn}/*"
       },